crash diagnostics: fix Windows write() size_t cast

+++ src/crash_handler.rs
@@ -48,7 +48,7 @@ pub fn append_to_log(msg: &str) {
                 libc::write(
                     fd,
                     line.as_ptr().add(written) as *const libc::c_void,
-                    line.len() - written,
+                    (line.len() - written) as libc::size_t,
                 )
             };
             if n <= 0 {
